In progress of solving APIHandler parsing of data.

+++ app/src/main/java/se/ju/agileandroidproject/Models/Coordinate.kt
@@ -3,5 +3,5 @@ package se.ju.agileandroidproject.Models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Coordinate(val lon: Double, val lat: Double) {
+data class Coordinate(val lon: Float, val lat: Float) {
 }
